Putting initial client structure in place
