+++ Cohort2/team4/client/src/components/layout.tsx
@@ -1,0 +1,13 @@
+import { Inter } from 'next/font/google'
+
+const inter = Inter({ subsets: ['latin'] })
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <>{children}</>
+  )
+}
