+++ Cohort2/team4/client/src/pages/_app.tsx
@@ -1,0 +1,12 @@
+import '@/styles/globals.css'
+import Layout from '../components/layout'
+
+import type { AppProps } from 'next/app'
+
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+			<Layout>
+				<Component {...pageProps} />
+			</Layout>
+	)
+}
