+++ Cohort2/team4/client/src/pages/index.tsx
@@ -1,0 +1,13 @@
+import Image from 'next/image'
+import Link from 'next/link'
+
+export default function Home() {
+  return (
+    <main className="flex min-h-screen flex-col items-center justify-between p-24">
+			<div>
+				<h1>TODO -- Replace with landing page</h1>
+				<Link href="events">Events Dashboard</Link>
+			</div>
+    </main>
+  )
+}
