+++ Cohort2/team4/client/src/pages/register.tsx
@@ -1,0 +1,3 @@
+export default function Register() {
+	return (<div>This is the registration page.</div>);
+}
